cleaning imports and moving sap and vm endpoints to right controllers
