+++ RFIDentity/src/main/java/com/rfidentity/RFIDentity/api/controller/VmItemController.java
@@ -1,13 +1,17 @@
 package com.rfidentity.RFIDentity.api.controller;
 
-import com.rfidentity.RFIDentity.api.dto.VmItemDTO;
+import com.rfidentity.RFIDentity.api.dto.ApiResponse;
+import com.rfidentity.RFIDentity.api.dto.DiffVmItemDTO;
 import com.rfidentity.RFIDentity.service.VmItemService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
 
 @RequestMapping("/api/vmItem")
 @RestController
@@ -16,8 +20,18 @@ public class VmItemController {
     @Autowired
     private VmItemService vmItemService;
 
-    @GetMapping
-    public List<VmItemDTO> getVmItem() {
-        return vmItemService.getAllVmItem();
+    @PutMapping("/updateVmItem/{inventoryId}/{assetId}")
+    public ResponseEntity<ApiResponse> updateVmItem(
+            @PathVariable Long inventoryId,
+            @PathVariable String assetId,
+            @RequestBody DiffVmItemDTO diffVmItemDTO) {
+
+        try{
+            String responseMessage = vmItemService.updateVmItem(assetId, inventoryId, diffVmItemDTO);
+            return new ResponseEntity<>(new ApiResponse(responseMessage), HttpStatus.OK);
+        }catch (Exception e){
+            return new ResponseEntity<>(new ApiResponse(e.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
 }
