+++ RFIDentity/src/main/java/com/rfidentity/RFIDentity/api/dto/InventoryDTO.java
@@ -1,6 +1,5 @@
 package com.rfidentity.RFIDentity.api.dto;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 
 import java.time.LocalDate;
