+++ RFIDentity/src/main/java/com/rfidentity/RFIDentity/api/dto/InventoryItemOutcomeDTO.java
@@ -1,7 +1,5 @@
 package com.rfidentity.RFIDentity.api.dto;
 
-import com.rfidentity.RFIDentity.model.InventoryItem;
-
 import lombok.Data;
 
 import java.time.LocalDate;
