+++ RFIDentity/src/main/java/com/rfidentity/RFIDentity/configuration/WebConfig.java
@@ -1,5 +1,4 @@
 package com.rfidentity.RFIDentity.configuration;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
